--- src/main/java/com/sparta/northwind/Category.java
@@ -1,22 +1,0 @@
-package com.sparta.northwind;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "categories")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "CategoryID", nullable = false)
-    private Integer id;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-//TODO [JPA Buddy] generate columns from DB
-}
