--- src/main/java/com/sparta/northwind/Supplier.java
@@ -1,22 +1,0 @@
-package com.sparta.northwind;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "suppliers")
-public class Supplier {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "SupplierID", nullable = false)
-    private Integer id;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-//TODO [JPA Buddy] generate columns from DB
-}
