+++ src/main/java/com/sparta/northwind/repositories/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.sparta.northwind.repositories;
+
+import com.sparta.northwind.entities.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+
+}
