+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
-
+spring.datasource.url=jdbc:mysql://localhost:3306/northwind
+spring.datasource.username=root
+spring.datasource.password=root
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+spring.jpa.show-sql=true
